[UT] Refractor Avro scanner UT
