+++ be/test/formats/avro/numeric_column_test.cpp
@@ -22,6 +22,7 @@
 #include <utility>
 
 #include "column/fixed_length_column.h"
+#include "exec/avro_test.h"
 #include "runtime/types.h"
 #include "util/defer_op.h"
 
@@ -36,13 +37,6 @@ extern "C" {
 namespace starrocks {
 
 class AvroAddNumericColumnTest : public ::testing::Test {};
-
-struct AvroHelper {
-    avro_schema_t schema = NULL;
-    avro_value_iface_t* iface = NULL;
-    avro_value_t avro_val;
-    std::string schema_text;
-};
 
 static void init_avro_value(std::string schema_path, AvroHelper& avro_helper) {
     std::ifstream infile_schema;
